+++ R/zzz.R
@@ -30,13 +30,13 @@
 }
 
 .onAttach <- function(...) {
-  if (!interactive() || stats::runif(1) > 0.25 || isTRUE(as.logical(Sys.getenv("AMR_silentstart", FALSE)))) {
+  if (!interactive() || stats::runif(1) > 0.1 || isTRUE(as.logical(Sys.getenv("AMR_silentstart", FALSE)))) {
     return()
   }
   packageStartupMessage("Thank you for using the AMR package! ",
                         "If you have a minute, please anonymously fill in this short questionnaire to improve the package and its functionalities:",
                         "\nhttps://msberends.github.io/AMR/survey.html",
-                        "\n[ permanently turn this message off with: Sys.setenv(AMR_silentstart = TRUE) ]")
+                        "\n[ prevent his notice with suppressPackageStartupMessages(library(AMR)) or use Sys.setenv(AMR_silentstart = TRUE) ]")
 }
 
 create_MO_lookup <- function() {